% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{bootstrap_mean_CI}
\alias{bootstrap_mean_CI}
\title{Combine results from list. Calculate p-value and AUC values}
\usage{
bootstrap_mean_CI(
  data,
  colName,
  percent = 0.95,
  method = c("percentile", "empirical"),
  num.boot
)
}
\arguments{
\item{data}{A data frame contains the interested numeric vector .}

\item{percent}{A number indicates the percentage of confidence interval.}

\item{method}{A character indicates the method used for computing bootstrap confidence interval. By default, this is set to `empirical`.}

\item{num.boot}{Number of bootstrap times.}
}
\value{
A data frame with lower and upper bootstrap confidence interval.
}
\description{
Combine results from list. Calculate p-value and AUC values
}
