% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{Boxplot_TBSig}
\alias{Boxplot_TBSig}
\title{Boxplot functions for list with inconsistent Signatures matched for each object
Also applies to consistent sigantures column names}
\usage{
Boxplot_TBSig(sig_list, sig_name, annotationName = "TBStatus")
}
\arguments{
\item{sig_list}{A list of data frame contains information with signature scores and Disease status}

\item{sig_name}{Name of the TB Signatures}

\item{annotationName}{Name of the disease column}
}
\value{
Boxplots
}
\description{
Boxplot functions for list with inconsistent Signatures matched for each object
Also applies to consistent sigantures column names
}
