% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotTBSig.R
\name{BoxplotTBSig}
\alias{BoxplotTBSig}
\alias{boxplotTBSig}
\title{Generate Boxplot for single gene signature scores across multiple studies}
\usage{
boxplotTBSig(object_list, gset, annotationColName)
}
\arguments{
\item{object_list}{A \code{list} of
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}
objects. Usually output from
\code{\link[TBSignatureProfiler]{runTBsigProfiler}}.}

\item{gset}{A character of vector of characters
represent name(s) of the signatures.
See \code{\link[TBSignatureProfiler]{TBsignatures}} for details.}

\item{annotationColName}{A character indicates the name of interest in the
object's column data.}
}
\value{
A \code{gtable} object that contains multiple Boxplots in the form of
  \code{ggplot} objects. Results show single signature's performance
  across multiple studies.
}
\description{
Generate Boxplot for single gene signature scores across multiple studies
}
\examples{
returned_resources <- curatedTBData(c("GSE107104", "GSE19435", "GSE19443"),
                                    dryrun = FALSE, curated.only = TRUE)
mysignatures <- list(Sweeney_OD_3 = c("DUSP3", "GBP5", "KLF2"))
re1 <- lapply(returned_resources, function(x)
                   subset_curatedTBData(x, "TBStatus", c("Control","PTB")))
re2 <- lapply(re1, function(x)
             TBSignatureProfiler::runTBsigProfiler(input = x,
                                                   useAssay = 1,
                                                   signatures = mysignatures,
                                                   algorithm = "ssGSEA",
                                                   update_genes = FALSE))
boxplotTBSig(object_list = re2,
             gset = "Sweeney_OD_3",
             annotationColName = "TBStatus")
}
