% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_curatedTBData.R
\name{check_annotation}
\alias{check_annotation}
\title{Check the annotation column name in the colData function}
\usage{
check_annotation(theObject, annotationColName)
}
\arguments{
\item{theObject}{A SummarizedExperiment/MultiAssayExperiment object.}

\item{annotationColName}{A character indicates feature of interest in the object's column data.}
}
\description{
Check the annotation column name in the colData function
}
