% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{combine_auc}
\alias{combine_auc}
\title{Combine results from list. Calculate p-value and AUC values}
\usage{
combine_auc(result_list, gset, annotationName = "TBStatus")
}
\arguments{
\item{result_list}{A list of SummarizedExperiment Object from `runTBsigProfiler`}
}
\value{
A data frame with signatures, p-value, and AUC
}
\description{
Combine results from list. Calculate p-value and AUC values
}
