% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineObjects.R
\name{combineObjects}
\alias{combineObjects}
\title{Merge samples with common genes from selected studies}
\usage{
combineObjects(object_list, experiment_name)
}
\arguments{
\item{object_list}{A \code{list} of \link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment} objects.
The assays contains object's assay contain expression data with probes mapped to gene symbol
\code{names(object_list)} should not be \code{NULL}}

\item{experiment_name}{A character/vector of character to choose the name of the assay from
\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment} object}
}
\value{
A \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment} object contains combined data from the input
}
\description{
Merge samples with common genes from selected studies
}
\examples{
geo <-  c("GSE19435", "GSE19439")
data_list <-  curatedTBData(geo, dryrun = FALSE, curated.only = TRUE)
sobject <- combineObjects(data_list, experiment_name = "assay_curated")
}
