% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSobject_Illumina4.R
\name{create_ColData}
\alias{create_ColData}
\title{Create Column data information from Illumina HumanHT-12 V3.0/4.0 expression beadchip and Affymetrix Human Genome U133 Plus 2.0 Array}
\usage{
create_ColData(geo_access, gse)
}
\arguments{
\item{geo_access}{A character that contains GEO accession number.}
}
\value{
A DataFrame object contains the patient information for each sample
}
\description{
Create Column data information from Illumina HumanHT-12 V3.0/4.0 expression beadchip and Affymetrix Human Genome U133 Plus 2.0 Array
}
\examples{
gse <- GEOquery::getGEO(geo_access, GSEMatrix =  F)
create_ColData_Illumina4("GSE39939",gse)

}
