% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSobject_Illumina4.R
\name{create_new_ColData}
\alias{create_new_ColData}
\title{Make column information consitent across dataset}
\usage{
create_new_ColData(col_data)
}
\value{
A DataFrame object contains the patient information for each sample
}
\description{
Make column information consitent across dataset
}
\examples{
gse <- GEOquery::getGEO(geo_access, GSEMatrix =  F)
col_data <- create_ColData("GSE39939",gse)
col_data_new <- create_new_ColData(col_data)
}
