% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSobject_Illumina3.R
\name{create_rawdata_Illumina3}
\alias{create_rawdata_Illumina3}
\alias{create_data_Illumina3}
\title{Create non-normalized data from Illumina HumanHT-12 V3.0 expression beadchip}
\usage{
create_data_Illumina3(Non_normalized_list)
}
\arguments{
\item{Non_normalized_list}{A list that contains non_normalized reads from each sample.}
}
\value{
A dataframe that contains non-normalized read counts.
}
\description{
Create non-normalized data from Illumina HumanHT-12 V3.0 expression beadchip
}
\examples{
Non_normalized_list <- download_data_Illumina3("GSE19442")
Non_normalized_data <- create_data_Illumina3(Non_normalized_lists)
}
