% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTBData.R
\name{create_RowData}
\alias{create_RowData}
\title{Create row data information from Illumina HumanHT-12 V3.0/4.0 expression beadchip and Affymetrix Human Genome U133 Plus 2.0 Array}
\usage{
create_RowData(dat_download, plat_access)
}
\arguments{
\item{dat_download}{A dataframe object read from `download_data_Illumina4` or `create_data_Illumina3`}

\item{plat_access}{A string that indicates platform access number}
}
\value{
A DataFrame object contains the information for each probe ID
}
\description{
Create row data information from Illumina HumanHT-12 V3.0/4.0 expression beadchip and Affymetrix Human Genome U133 Plus 2.0 Array
}
\examples{
Non_normalized_data <- download_data_Illumina4("GSE19442")
plat_access <- "GPL10558"
row_data <- create_RowData(Non_normalized_data,"GPL10558")
}
