% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTBData.R
\name{create_standard_ColData}
\alias{create_standard_ColData}
\title{Make column information consitent across dataset}
\usage{
create_standard_ColData(col_data)
}
\arguments{
\item{col_data}{Column data information created from `create_ColData`}
}
\value{
A DataFrame object contains the patient information for each sample
}
\description{
Make column information consitent across dataset
}
\examples{
gse <- GEOquery::getGEO(geo_access, GSEMatrix =  F)
col_data <- create_ColData("GSE39939",gse)
col_data_new <- create_standard_ColData(col_data)
}
