% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_curatedTBData.R
\name{CreateObject}
\alias{CreateObject}
\alias{CreateObject,Sobject-method}
\alias{CreateObject,Mobject-method}
\title{Create SummarizedExperiment/MultiAssayExperiment object}
\usage{
CreateObject(theObject, ...)

\S4method{CreateObject}{Sobject}(theObject)

\S4method{CreateObject}{Mobject}(theObject, assay_type = "assay_reprocess")
}
\arguments{
\item{theObject}{A class either Sobject or Mobject}

\item{...}{Extra named arguments passed to function}
}
\description{
Create SummarizedExperiment/MultiAssayExperiment object
}
