% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_curatedTBData.R
\name{curatedTBData}
\alias{curatedTBData}
\alias{curatedTBData,character-method}
\title{Create curated Tuberculosis transcriptome data from sequencing platform: GPL6947/GPL10558/GPL570}
\usage{
curatedTBData(geo_access, ...)

\S4method{curatedTBData}{character}(geo_access, plat_access = c("GPL6947", "GPL10558", "GPL570"))
}
\arguments{
\item{geo_access}{A character stands for GEO accession number of the dataset}

\item{...}{Extra named arguments passed to function}

\item{plat_access}{A character stands for acession number of squencing platform}
}
\description{
Create curated Tuberculosis transcriptome data from sequencing platform: GPL6947/GPL10558/GPL570
}
