% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTBData.R
\name{curatedTBData}
\alias{curatedTBData}
\title{Import curatedTBData
A function to load available curatedTBData from the package.}
\usage{
curatedTBData(geoAccession)
}
\arguments{
\item{geoAccession}{A character/vector that contains geo accession information.
If `geoAccession == All``, load all avaible studies in the package.}
}
\value{
A MultiAssayExperiment Object correpsonding to the `geoAccession`
}
\description{
Import curatedTBData
A function to load available curatedTBData from the package.
}
