% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_curatedTBData.R
\name{.affy2}
\alias{.affy2}
\title{Affymetrix Human Genome U133 Plus 2.0 Array}
\usage{
.affy2(geo_access, plat_access)
}
\arguments{
\item{geo_access}{A character stands for GEO accession number of the dataset}

\item{plat_access}{A character stands for acession number of squencing platform}
}
\description{
Affymetrix Human Genome U133 Plus 2.0 Array
}
