% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_AUC.R
\name{.bootstrap_mean_CI}
\alias{.bootstrap_mean_CI}
\title{Compute bootstrapped confidence interval for signature's mean AUC across studies}
\usage{
.bootstrap_mean_CI(
  df,
  colName,
  method = c("percentile", "empirical"),
  num.boot,
  percent = 0.95
)
}
\arguments{
\item{df}{A data frame/matrix contains the interested numeric vector obtained
from \code{combine_auc}. See \code{\link[curatedTBData]{combine_auc}} for details.}

\item{colName}{A character string specifying the column name of the data frame
for bootstrapping.}

\item{method}{A character string specifying the method used for computing bootstrap confidence interval.
The choices are c("percentile","empirical"). The default is "empirical".}

\item{num.boot}{Numeric. Number of bootstrap times. Default is `100`.}

\item{percent}{A number indicates the percentage of confidence interval.
Default is `0.95`.}
}
\value{
A data frame contains mean AUC, lower and upper bootstrap confidence interval
for single gene signature across multiple studies.
}
\description{
Compute bootstrapped confidence interval for signature's mean AUC across studies
}
