% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_auc.R
\name{.check_input}
\alias{.check_input}
\title{Check the input list format}
\usage{
.check_input(object_list)
}
\arguments{
\item{object_list}{A \code{list} of
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
}
\value{
Nothing is returned for this function.
}
\description{
Function to check whether input list has desired format.
}
