% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTBData.R
\name{.createMultiAssayExperimentObject}
\alias{.createMultiAssayExperimentObject}
\title{Create MultiAssayExperiment Object
A function to to create MultiAssayExperiment Object given geo accession number
from the curatedTBData package.}
\usage{
.createMultiAssayExperimentObject(geoAccession, fileNamesSelected)
}
\arguments{
\item{geoAccession}{A character/vector that contains geo accession information.
If `geoAccession == All``, load all avaible studies in the package.}

\item{fileNameSelected}{A character for slected file path.}
}
\value{
A list of MultiAssayExperiment objects
}
\description{
Create MultiAssayExperiment Object
A function to to create MultiAssayExperiment Object given geo accession number
from the curatedTBData package.
}
