% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTBData.R
\name{.curatedTBDataLoadData}
\alias{.curatedTBDataLoadData}
\title{Load selected curatedTBData
A function to load selected datasets from the package.}
\usage{
.curatedTBDataLoadData(geoAccession, loadAll)
}
\arguments{
\item{geoAccession}{A character/vector that contains geo accession information.
If `geoAccession == All``, load all avaible studies in the package.}

\item{loadAll}{Logical. An logical value to indicate whether load all available studies.}
}
\value{
A list of MultiAssayExperiment objects
}
\description{
Load selected curatedTBData
A function to load selected datasets from the package.
}
