% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTBData.R
\name{.illumina4}
\alias{.illumina4}
\title{Functions to create assays and row data from Illumina HumanHT-12 V4.0 expression beadchip}
\usage{
.illumina4(geo_access, plat_access)
}
\arguments{
\item{geo_access}{A character stands for GEO accession number of the dataset}

\item{plat_access}{A character stands for acession number of squencing platform}
}
\description{
Functions to create assays and row data from Illumina HumanHT-12 V4.0 expression beadchip
}
