% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotTBSig.R
\name{.signature_filter}
\alias{.signature_filter}
\title{Subset signatures scores and disease status from column data of SummarizedExperiment objects.}
\usage{
.signature_filter(object_list, gset, annotationColName)
}
\arguments{
\item{object_list}{A `list` of [SummarizedExperiment][SummarizedExperiment::SummarizedExperiment-class] objects.
Usually output from \code{\link[TBSignatureProfiler]{runTBsigProfiler}}.}

\item{gset}{A character indicates the name of the signatures.}

\item{annotationColName}{A character indicates the name of interest in the object's column data.}
}
\value{
A `list` of `data.frame` contains the `annotationColName`, prediction score and study name of single object.
}
\description{
Subset signatures scores and disease status from column data of SummarizedExperiment objects.
}
