% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSobject_Affy.R
\name{download_data_Affy}
\alias{download_data_Affy}
\title{Download data from Illumina HumanHT-12 V4.0 expression beadchip.}
\usage{
download_data_Affy(geo_access)
}
\arguments{
\item{geo_access}{A character that contains GEO accession number.}
}
\value{
A dataframe that contains non-normalized read counts.
}
\description{
Download data from Illumina HumanHT-12 V4.0 expression beadchip.
}
\examples{
download_data_Affy("GSE36238")
}
