% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_curatedTBData.R
\name{download_data_Affy2}
\alias{download_data_Affy2}
\title{Download data from Affymetrix Human Genome U133 Plus 2.0 Array.}
\usage{
download_data_Affy2(geo_access)
}
\arguments{
\item{geo_access}{A character that contains GEO accession number.}
}
\value{
A dataframe that contains non-normalized read counts.
}
\description{
Download data from Affymetrix Human Genome U133 Plus 2.0 Array.
}
\examples{
download_data_Affy2("GSE36238")
}
