% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{get_auc_distribution}
\alias{get_auc_distribution}
\title{Obtain ridge plots for emprirical AUC distribution for signature scores.}
\usage{
get_auc_distribution(result_list)
}
\value{
Ridge plot with median line
}
\description{
Obtain ridge plots for emprirical AUC distribution for signature scores.
}
\examples{
mobject <- get_expr_symbol(sobject)
p_ridge <- get_auc_distribution(result_list,"Disease",signatureColNames)
}
