% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{get_auc_stats}
\alias{get_auc_stats}
\title{Obtain pvalue, emprirical AUC, and Bootstrap Confidence Interval for each signature.}
\usage{
get_auc_stats(
  SE_scored,
  annotationColName = "TBStatus",
  signatureColNames,
  num.boot = NULL,
  percent = 0.95,
  AUC.abs = FALSE,
  BPPARAM = BiocParallel::SerialParam(progressbar = TRUE)
)
}
\arguments{
\item{SE_scored}{A \code{SummarizedExperiment} Object from TB signature profiling.}

\item{annotationColName}{A character indicates feature of interest in the object's column data}

\item{signatureColNames}{A character/vector contains name of gene signature.}

\item{num.boot}{Number of bootstrapping.}

\item{percent}{A number speciying the percentage of the confidence interval.}

\item{AUC.abs}{Boolean. If AUC.abs = TRUE, return the AUC values from function \code{\link[ROCit]{rocit}}.
If AUC.abs = FALSE, return the AUC values for max(AUC, 1-AUC).}

\item{BPPARAM}{An instance inherited from \code{bplappy}.
See \code{\link[BiocParallel]{bplapply}} for details.}
}
\value{
A data frame/datatable contains p-value from two-sample t-test and AUC value for each signature.
}
\description{
Obtain pvalue, emprirical AUC, and Bootstrap Confidence Interval for each signature.
}
