% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{get_boxplot}
\alias{get_boxplot}
\title{Get boxplot plot comparison across all signatures}
\usage{
get_boxplot(result_list, name)
}
\arguments{
\item{result_list}{A list that contains TB signature scores for each dataset}

\item{name}{A string for signature name}
}
\value{
Combined boxplot that shows specific signature across all dataset
}
\description{
Get boxplot plot comparison across all signatures
}
\examples{
p_boxplot <- get_boxplot(TB_sobject,"Sloot_HIV_2")
}
