% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTBData.R
\name{get_curatedTBData}
\alias{get_curatedTBData}
\title{Combine individual data to SummarizedExperiment/MultiAssayExperiment object}
\usage{
get_curatedTBData(geo_access)
}
\arguments{
\item{geo_access}{A character/vector that contains geo accession number. If NULL, get all avaible studies.}
}
\value{
A list of SummarizedExperiment/MultiAssayExperiment objects
}
\description{
Combine individual data to SummarizedExperiment/MultiAssayExperiment object
}
