% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_rank.R
\name{get_gsea_rank}
\alias{get_gsea_rank}
\title{Obtain signature gene rank}
\usage{
get_gsea_rank(data_sobject, TBSig, annotationColNames = "TBStatus")
}
\arguments{
\item{data_sobject}{A SummarizedExperiment Object contains assays and column data with disease status.}

\item{annotationColNames}{A character indicates feature of interest in the object's column data.}

\item{TBsignatures}{A list of signature genes with list names as signature name.}
}
\value{
A list of dataframes (single dataframe if only one signature) with Disease status and signature genes rank.
}
\description{
Obtain signature gene rank
}
\examples{
data_sobject <- TBSignatureProfiler::TB_hiv
TBSig <- TBSignatureProfiler::TBsignatures[c("Sweeney_OD_3","Roe_3")]
annotationColNames <- "Disease"
get_gsea_rank(data_sobject,TBSig,annotationColNames)

}
