% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_rank.R
\name{get_rank_boxplot}
\alias{get_rank_boxplot}
\title{Obtain boxplot of gene rank boxplot for each signature}
\usage{
get_rank_boxplot(sig_list, gset)
}
\arguments{
\item{sig_list}{A (nested) list of data frame}

\item{gset}{A character indicates the name of the signature}
}
\value{
A group of boxplots show signature gene ranks under difference disease status.
}
\description{
Obtain boxplot of gene rank boxplot for each signature
}
\examples{
data_sobject <- TBSignatureProfiler::TB_hiv
TBSig <- TBSignatureProfiler::TBsignatures[c("Sweeney_OD_3","Roe_3")]
annotationColNames <- "Disease"
sig_list <- get_gsea_rank(data_sobject,TBSig,annotationColNames)
get_rank_boxplot(sig_list,"Sweeney_OD_3")
}
