% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{get_sobject_TBSig}
\alias{get_sobject_TBSig}
\title{Create sobjects for TB signature profiling based on specific TB status.}
\usage{
get_sobject_TBSig(multi_object, disease1, disease2)
}
\arguments{
\item{multi_object}{an multi-assay summarized experiment}

\item{disease1}{A character chooses from "Control", "Latent", "PTB", "OD"}

\item{disease2}{A character chooses from "Control", "Latent", "PTB", "OD".
TBStatus1 and TBStatus2 should be different from each other}
}
\value{
A summarized experiment object for TB signature profiling
}
\description{
Create sobjects for TB signature profiling based on specific TB status.
}
\examples{
sobject <- dat("GSE39939_sobject")
mobject <- Create_MultiAssay_object(sobject)
sobject_TBSig <- get_sobject_TBSig(mobject,"PTB","Latent")
}
