% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchProbe.R
\name{MatchProbe}
\alias{MatchProbe}
\alias{MatchProbe,SummarizedExperiment-method}
\alias{MatchProbe,MultiAssayExperiment-method}
\title{S4 method matches probeID to gene symbol by creating MultiAssayExperiement
object from either SummarizedExperiment or MultiAssayExperiment Object.}
\usage{
MatchProbe(theObject, ...)

\S4method{MatchProbe}{SummarizedExperiment}(
  theObject,
  useAssay,
  FUN = median,
  createExperimentName = "assay_MatchProbe",
  only.matrix = FALSE
)

\S4method{MatchProbe}{MultiAssayExperiment}(
  theObject,
  useAssay,
  FUN = median,
  createExperimentName = "assay_MatchProbe",
  only.matrix = FALSE
)
}
\arguments{
\item{theObject}{A SummarizedExperiment/MultiAssayExperiment Object.}

\item{...}{Extra named arguments passed to function.}

\item{useAssay}{A character indicats the assay names (partial name) of the
SummarizedExperiment Object.}

\item{createExperimentName}{A character specifying the names of the new
experiment matrix.}

\item{only.matrix}{\code{boolean}. A logical value indicates whether return the
processed matrix only.
The default is FALSE. TRUE for only output the matched matrix.}
}
\description{
S4 method matches probeID to gene symbol by creating MultiAssayExperiement
object from either SummarizedExperiment or MultiAssayExperiment Object.
}
