% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchProbe.R
\name{MatchProbe}
\alias{MatchProbe}
\title{Match ProbeID to gene symbol by creating MultiAssayExperiement object from SummarizedExperiment/MultiAssayExperiment Object.}
\usage{
MatchProbe(sobject)
}
\arguments{
\item{sobject}{A summarized experiment object}
}
\value{
A MultiAssay Experiment Object contains 1. Input Summarized Experiment Object. 2. read count matrix collapse by gene symbol
}
\description{
Match ProbeID to gene symbol by creating MultiAssayExperiement object from SummarizedExperiment/MultiAssayExperiment Object.
}
\examples{
dat("GSE39939_sobject")
MatchProbe(GSE39939_sobject)
}
