% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTBData.R
\name{MatchSampleID}
\alias{MatchSampleID}
\title{Match Description ID to Sample ID in some Illumina 4.0 datasets}
\usage{
MatchSampleID()
}
\description{
Match Description ID to Sample ID in some Illumina 4.0 datasets
}
