% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDP.R
\name{MDP}
\alias{MDP}
\alias{MDP,SummarizedExperiment,NULL-method}
\alias{MDP,MultiAssayExperiment,NULL-method}
\alias{MDP,SummarizedExperiment,list-method}
\alias{MDP,MultiAssayExperiment,list-method}
\title{S4 Method calcualtes molecular degree of perturbation scores for SummarizedExperiment/MultiAssayExperiment object.}
\usage{
MDP(theObject, gset, ...)

\S4method{MDP}{SummarizedExperiment,`NULL`}(theObject, gset = NULL, assay_name, GSE = GSE)

\S4method{MDP}{MultiAssayExperiment,`NULL`}(
  theObject,
  gset = NULL,
  experiment_name = "assay_reduce",
  assay_name = 1,
  GSE = GSE
)

\S4method{MDP}{SummarizedExperiment,list}(theObject, gset = gset, assay_name = 1)

\S4method{MDP}{MultiAssayExperiment,list}(
  theObject,
  gset = gset,
  experiment_name = "assay_reduce",
  assay_name = 1,
  GSE = GSE
)
}
\arguments{
\item{theObject}{A SummarizedExperiment/MultiAssayExperiment object.}

\item{gset}{A vector contians name(s) of the signatures.}

\item{...}{Extra named arguments passed to function.}

\item{assay_name}{A number/character specifying the assay name of the SummarizedExperiment objectt.}

\item{GSE}{List of character strings specifying the name of the study, usually GEO accession number.}

\item{experiment_name}{A character indicates the name of the experiment within MultiAssayExperiment object.}
}
\description{
S4 Method calcualtes molecular degree of perturbation scores for SummarizedExperiment/MultiAssayExperiment object.
}
