% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDP.R
\name{MDP}
\alias{MDP}
\title{calculate molecular degree of perturbation scores for selected sample}
\usage{
MDP(theObject, gset, ...)
}
\description{
calculate molecular degree of perturbation scores for selected sample
}
