% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_curatedTBData.R
\docType{class}
\name{Mobject}
\alias{Mobject}
\title{Create MultiAssay Experiment object}
\description{
Create MultiAssay Experiment object
}
\section{Slots}{

\describe{
\item{\code{assay_reprocess}}{A matrix contatins gene expression data}

\item{\code{assay_raw}}{A matrix contatins gene expression data}

\item{\code{row_data}}{A DataFrame class contains gene expression data with different dimensions}

\item{\code{primary}}{A DataFrame class contains sample information}

\item{\code{meta_data}}{A MIAME class contains experiment information}
}}

