% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{Normalization}
\alias{Normalization}
\alias{Normalization,SummarizedExperiment-method}
\alias{Normalization,MultiAssayExperiment-method}
\title{Normalization for microarray and RNA-seq transcriptome data.}
\usage{
Normalization(theObject, ...)

\S4method{Normalization}{SummarizedExperiment}(theObject, method = "quantile", ...)

\S4method{Normalization}{MultiAssayExperiment}(
  theObject,
  experiment_type = c("assay_raw", "assay_reprocess"),
  method = "TMM"
)
}
\arguments{
\item{theObject}{A SummarizedExperiment/MultiAssayExperiment object.}

\item{...}{Extra named arguments passed to function.}

\item{method}{character string specifying the normalization method to be used. See `limma::normalizeBetweenArrays` for microarray, and `edgeR::TMM` for RNA sequence data.}

\item{experiment_type}{A character indicates the name of the experiment within MultiAssayExperiment object.}
}
\description{
Normalization for microarray and RNA-seq transcriptome data.
}
