% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{Normalization}
\alias{Normalization}
\title{Normalization for microarray and RNA-seq transcriptome data.}
\usage{
Normalization(theObject, ...)
}
\description{
Normalization for microarray and RNA-seq transcriptome data.
}
