% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{NormalizeReads}
\alias{NormalizeReads}
\title{Normalization for microarray and RNA-seq transcriptome data.}
\usage{
NormalizeReads(theObject, ...)
}
\description{
Normalization for microarray and RNA-seq transcriptome data.
}
