% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeplot_auc.R
\name{ridgeplot_auc}
\alias{ridgeplot_auc}
\title{Obtain ridge plots for empirical AUC distribution for signature scores.}
\usage{
ridgeplot_auc(aucs_result)
}
\arguments{
\item{aucs_result}{A `data.frame` contains signatures, p-value, and AUC.
Usually the output from \code{\link{combine_auc}}.}
}
\value{
Ridge plot with median line
}
\description{
Obtain ridge plots for empirical AUC distribution for signature scores.
}
\examples{
aucs_result <- data.frame(Signature = c("Anderson_42", "Anderson_OD_51", "Berry_393"),
                          AUC = stats::runif(3,0,0.5))
p_ridge <- get_auc_distribution(aucs_result)
}
