% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_curatedTBData.R
\name{SCAN_reprocess_FALSE}
\alias{SCAN_reprocess_FALSE}
\title{Combine individual data into SummarizedExperiment object}
\usage{
SCAN_reprocess_FALSE(geo_access, include.SCAN)
}
\arguments{
\item{geo_access}{A character/vector that contains geo accession number. If All, get all avaible studies.}

\item{include.SCAN}{A logical value indicates whether include normalized data processed by SCAN into the final output.
The default in FALSE.}
}
\description{
\code{SCAN_reprocess_FALSE()} combines individual data into SummarizedExperiment
object for all available study.
This function will not include reprocessed RNA-seq information.
}
\examples{
SCAN_reprocess_TRUE(geo_access = "GSE39939", include.SCAN = TRUE)
SCAN_reprocess_TRUE(geo_access = "GSE101705", include.SCAN = TRUE)
}
