% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_curatedTBData.R
\name{SCAN_reprocess_TRUE}
\alias{SCAN_reprocess_TRUE}
\title{Combine individual data to SummarizedExperiment/MultiAssayExperiment object
when include.reprocess = TRUE}
\usage{
SCAN_reprocess_TRUE(geo_access, include.SCAN)
}
\arguments{
\item{geo_access}{A character/vector that contains geo accession number. If All, get all avaible studies.}

\item{include.SCAN}{A logical value indicates whether include normalized data processed by SCAN into the final output when available.
The default in FALSE.}
}
\description{
Combine individual data to SummarizedExperiment/MultiAssayExperiment object
when include.reprocess = TRUE
}
