% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{SignatureFilter}
\alias{SignatureFilter}
\title{Get signatures with only siganture scores and disease status}
\usage{
SignatureFilter(sig_list, gset, ...)
}
\description{
Get signatures with only siganture scores and disease status
}
