% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_curatedTBData.R
\docType{class}
\name{Sobject-class}
\alias{Sobject-class}
\alias{Sobject}
\title{Sobject Class
Create SummarizedExperiment object for curatedTBData}
\description{
Sobject Class
Create SummarizedExperiment object for curatedTBData
}
\section{Slots}{

\describe{
\item{\code{assay}}{A matrix contatins gene expression data}

\item{\code{row_data}}{A DataFrame class contains gene information}

\item{\code{column_data}}{A DataFrame class contains sample information}

\item{\code{meta_data}}{A MIAME class contains experiment information}
}}

