% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsetting.R
\name{Subset_curatedTBData}
\alias{Subset_curatedTBData}
\alias{Subset_curatedTBData,SummarizedExperiment-method}
\alias{Subset_curatedTBData,MultiAssayExperiment-method}
\title{Subsetting objects based on single/multiple conditions}
\usage{
Subset_curatedTBData(theObject, ...)

\S4method{Subset_curatedTBData}{SummarizedExperiment}(
  theObject,
  annotationColName,
  annotationCondition,
  UseAssay,
  ...
)

\S4method{Subset_curatedTBData}{MultiAssayExperiment}(
  theObject,
  annotationColName,
  annotationCondition,
  experiment_type
)
}
\arguments{
\item{theObject}{A SummarizedExperiment/MultiAssayExperiment object.}

\item{...}{Extra named arguments passed to function.}

\item{annotationColName}{A character indicates feature of interest in the object's column data.}

\item{annotationCondition}{A vector indicates conditions want to be subsetted.}

\item{experiment_type}{A character indicates the name of the experiment within MultiAssayExperiment object.
Expect theObject[[experiment_type]] to be a matrix. Two special cases are:
When experiment_type is "all". Perform whole MultiAssayExperiment subsetting, output is in the form of MultiAsaayExperiment object.
When experiment_type is "assay_raw". Perform subsetting on the SummarizedExperiment Object.}
}
\description{
Subsetting objects based on single/multiple conditions
}
