% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsetting.R
\name{SubsetDisease}
\alias{SubsetDisease}
\title{Subsetting objects based on single/multiple conditions}
\usage{
SubsetDisease(theObject, ...)
}
\description{
Subsetting objects based on single/multiple conditions
}
