% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsetting.R
\name{SubsetSample}
\alias{SubsetSample}
\title{Subsetting objects based on single/multiple conditions}
\usage{
SubsetSample(theObject, ...)
}
\description{
Subsetting objects based on single/multiple conditions
}
