% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsetting.R
\name{SubsetTBStatus}
\alias{SubsetTBStatus}
\alias{SubsetTBStatus,SummarizedExperiment-method}
\alias{SubsetTBStatus,MultiAssayExperiment-method}
\title{Subsetting objects based on single/multiple conditions}
\usage{
SubsetTBStatus(theObject, ...)

\S4method{SubsetTBStatus}{SummarizedExperiment}(theObject, annotationColName, diseases, ...)

\S4method{SubsetTBStatus}{MultiAssayExperiment}(
  theObject,
  annotationColName,
  diseases,
  experiment_type = c("all", "assay_raw", "assay_reprocess", "assay_reduce")
)
}
\arguments{
\item{theObject}{A SummarizedExperiment/MultiAssayExperiment object.}

\item{...}{Extra named arguments passed to function.}

\item{annotationColName}{A character indicates feature of interest in the object's column data.}

\item{diseases}{A vector indicates (disease)conditions want to subset.}

\item{experiment_type}{A character indicates the name of the experiment within MultiAssayExperiment object.
Choices for experiment_type are: "all", "assay_raw", "assay_reprocess", "assay_reduce".
When experiment_type is "all". Perform whole MultiAssayExperiment subsetting, output is in the form of MultiAsaayExperiment object.
When experiment_type is "assay_reduce" or "assay_reprocess". Perform subsetting on matrix with experiment name "assay_reduce"/"assay_reprocess" respectively, output is in the form of SummarizedExperiment object.}
}
\description{
Subsetting objects based on single/multiple conditions
}
