% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsetting.R
\name{SubsetTBStatus}
\alias{SubsetTBStatus}
\alias{SubsetTBStatus,SummarizedExperiment-method}
\alias{SubsetTBStatus,MultiAssayExperiment-method}
\title{Subsetting objects based on single/multiple conditions
Output should be in the orginal format either summarizedExperiment or MultiAlssayExperiment
experiment=NULL in default
Output is in the form of SummarizedExperiment format for TBsignatureProfier
through indication of which assay you are going to work with}
\usage{
SubsetTBStatus(theObject, ...)

\S4method{SubsetTBStatus}{SummarizedExperiment}(theObject, annotationColName, diseases, ...)

\S4method{SubsetTBStatus}{MultiAssayExperiment}(
  theObject,
  annotationColName,
  diseases,
  experiment_type = c("all", "assay_raw", "assay_reprocess", "assay_reduce")
)
}
\arguments{
\item{theObject}{A SummarizedExperiment/MultiAssayExperiment object}

\item{...}{Extra named arguments passed to function}

\item{annotationColName}{A character indicates feature of interest in the object's column data}

\item{diseases}{A vector indicates conditions want to subset}

\item{experiment_type}{A character indicates the name of the experiment within MultiAssayExperiment object}
}
\description{
Subsetting objects based on single/multiple conditions
Output should be in the orginal format either summarizedExperiment or MultiAlssayExperiment
experiment=NULL in default
Output is in the form of SummarizedExperiment format for TBsignatureProfier
through indication of which assay you are going to work with
}
