% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{TBSig_multi_object}
\alias{TBSig_multi_object}
\title{TB Signature Profiling for multi assay object}
\usage{
TBSig_multi_object(multi_object, disease1, disease2)
}
\arguments{
\item{multi_object}{A MultiAssay Experiment Object}

\item{disease1}{A string indicating TB status, choose from c("PTB","Latent","Control","OD")}

\item{disease2}{A string indicating TB status, choose from c("PTB","Latent","Control","OD"),`disease1` should be different from `disease2`.}
}
\value{
A MultiAssay Experiment Object contains 1. Input Summarized Experiment Object. 2. read count matrix collapse by gene symbol
}
\description{
TB Signature Profiling for multi assay object
}
\examples{
dat("GSE39939_sobject")
multi_object <- Create_MultiAssay_object(GSE39939_sobject)
TB_sobject <- TBSig_multi_object(multi_object)
}
