% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineObjects.R
\name{update_gene_symbol}
\alias{update_gene_symbol}
\title{Update gene names from input data}
\usage{
update_gene_symbol(dat_exprs)
}
\arguments{
\item{dat_exprs}{A \code{data.frame} with row names being gene symbol to be updated}
}
\value{
A \code{data.frame} with updated gene symbol as row names
}
\description{
Update gene names from input data
}
