% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotTBSig.R
\name{BoxplotTBSig}
\alias{BoxplotTBSig}
\alias{boxplotTBSig}
\title{Boxplot functions for list of signature scores across studies
Also applies to consistent signature column names.}
\usage{
boxplotTBSig(object_list, gset, annotationColName = "TBStatus")
}
\arguments{
\item{object_list}{A `list` of [SummarizedExperiment][SummarizedExperiment::SummarizedExperiment-class] objects.
Usually output from \code{\link[TBSignatureProfiler]{runTBsigProfiler}}}

\item{gset}{A character of vector of characters represent name(s) of the signatures.
See \code{\link[TBSignatureProfiler]{TBsignatures}} for details.}

\item{annotationColName}{A character indicates the name of interest in the object's column data.}
}
\value{
A `gtable` contains single signature's performance across multiple studies.
}
\description{
Boxplot functions for list of signature scores across studies
Also applies to consistent signature column names.
}
