% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{BoxplotTBSig}
\alias{BoxplotTBSig}
\title{Boxplot functions for list with Signatures matched for each object (particularly, for inconsistent mattching)
Also applies to consistent sigantures column names}
\usage{
BoxplotTBSig(sig_list, sig_name, ...)
}
\description{
Boxplot functions for list with Signatures matched for each object (particularly, for inconsistent mattching)
Also applies to consistent sigantures column names
}
