% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotTBSig.R
\name{BoxplotTBSig}
\alias{BoxplotTBSig}
\alias{boxplotTBSig}
\title{Boxplot functions for list of signature scores across studies
Also applies to consistent signature column names.}
\usage{
boxplotTBSig(object_list, gset, annotationColName = "TBStatus")
}
\arguments{
\item{object_list}{A `list` of [SummarizedExperiment][SummarizedExperiment::SummarizedExperiment-class] objects.
Usually output from \code{\link[TBSignatureProfiler]{runTBsigProfiler}}}

\item{gset}{A character of vector of characters represent name(s) of the signatures.
See \code{\link[TBSignatureProfiler]{TBsignatures}} for details.}

\item{annotationColName}{A character indicates the name of interest in the object's column data.}
}
\value{
A `gtable` object that contains multiple boxplots in the form of `ggplot` objects.
Results show single signature's performance across multiple studies.
}
\description{
Boxplot functions for list of signature scores across studies
Also applies to consistent signature column names.
}
\examples{
returned_resources <- curatedTBData(c("GSE107104", "GSE19435", "GSE19443"),
                                    dryrun = FALSE, curated.only = TRUE)
mysignatures <- list(Sweeney_OD_3 = c("DUSP3", "GBP5", "KLF2"))
re1 <- lapply(returned_resources, function(x)
                   subset_curatedTBData(x, annotationColName = "TBStatus",
                                        annotationCondition = c("Control","PTB")))
re2 <- lapply(re1, function(x)
               TBSignatureProfiler::runTBsigProfiler(input = x,
                                                     useAssay = 1,
                                                     signatures = mysignatures,
                                                     algorithm = "ssGSEA",
                                                     update_genes = FALSE))
boxplotTBSig(object_list = re2, gset = "Sweeney_OD_3", annotationColName = "TBStatus")
}
