% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsetting.R
\name{CombineObjects}
\alias{CombineObjects}
\title{Combine samples with common genes from selected studies, usually run after `MatchProbe`}
\usage{
CombineObjects(object_list, gse_name = NULL, experiment_type = "assay_reduce")
}
\arguments{
\item{object_list}{A list contains expression data with mapped gene symbol.}

\item{gse_name}{A charac/vector (GEO accession number) contains object name want to combine.}
}
\value{
A SummarizedExperiment Object contains combined data from several objects.
}
\description{
Combine samples with common genes from selected studies, usually run after `MatchProbe`
}
