% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_curatedTBData.R
\name{CombineObjects}
\alias{CombineObjects}
\title{Combine samples with common genes from selected studies,
usually run after \code{\link{MatchProbe}}}
\usage{
CombineObjects(
  object_list,
  list_name = NULL,
  experiment_name,
  batch.adjust = TRUE
)
}
\arguments{
\item{object_list}{A list contains expression data with probes mapped to gene symbol.}

\item{list_name}{A character/vector contains object name to be selected to merge.}

\item{experiment_name}{A character/vector to choose the name of the experiment from MultiAssayExperiment Object.}

\item{batch.adjust}{A logical value indicating whether adjust for the batch effect.
Default is TRUE.}
}
\value{
A SummarizedExperiment Object contains combined data from several objects.
}
\description{
Combine samples with common genes from selected studies,
usually run after \code{\link{MatchProbe}}
}
\examples{
list_name <-  c("GSE101705","GSE54992","GSE19444")
data_list <-  get_curatedTBData(list_name)
object_norm <- lapply(data_list, function(x)
                               Normalization(x, microarray_method = "quantile",
                               RNAseq_method = "TMM", experiment_name = "assay_raw"))
object_match <- lapply(object_norm, function(x)
                               MatchProbe(x, UseAssay = c("TMM","quantile","RMA"),
                               createExperimentName = "assay_MatchProbe"))
sobject <- CombineObjects(object_match, list_name,
                          experiment_name = "assay_MatchProbe", batch.adjust = TRUE)
}
