% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_curatedTBData.R
\name{CombineObjects}
\alias{CombineObjects}
\title{Combine samples with common genes from selected studies,
usually run after matching prob sets to gene symbol. See \code{\link{MatchProbe}}}
\usage{
CombineObjects(
  object_list,
  list_name = NULL,
  experiment_name = NULL,
  useAssay = NULL
)
}
\arguments{
\item{object_list}{A list contains expression data with probes mapped to gene symbol.}

\item{list_name}{A character/vector contains object name to be selected to merge.}

\item{experiment_name}{A character/vector to choose the name of the experiment from MultiAssayExperiment Object.}
}
\value{
A SummarizedExperiment Object contains combined data from several studies.
}
\description{
Combine samples with common genes from selected studies,
usually run after matching prob sets to gene symbol. See \code{\link{MatchProbe}}
}
\examples{
list_name <-  c("GSE101705","GSE54992","GSE19444")
data_list <-  get_curatedTBData(list_name)
sobject <- CombineObjects(data_list, list_name,
                          experiment_name = "assay_curated")
}
