% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_curatedTBData.R
\name{combineObjects}
\alias{combineObjects}
\title{Merge samples with common genes from selected studies}
\usage{
combineObjects(object_list, experiment_name = NULL)
}
\arguments{
\item{object_list}{A `list` of [MultiAssayExperiment][MultiAssayExperiment::MultiAssayExperiment-class] objects.
The assays contains object's assay contain expression data with probes mapped to gene symbol.
`names(object_list)` should NOT be `NULL`.}

\item{experiment_name}{A character/vector of character to choose the name of the assay from
[MultiAssayExperiment][MultiAssayExperiment::MultiAssayExperiment-class] object.}
}
\value{
A [SummarizedExperiment][SummarizedExperiment::SummarizedExperiment-class]
object contains combined data from the input.
}
\description{
Merge samples with common genes from selected studies
}
\examples{
geo <-  c("GSE19435", "GSE19439")
data_list <-  curatedTBData(geo, dryrun = FALSE, curated.only = TRUE)
sobject <- combineObjects(data_list, experiment_name = "assay_curated")
}
