% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{CombineObjects}
\alias{CombineObjects}
\title{Combine samples with common genes from selected objects}
\usage{
CombineObjects(object_list, gse_name)
}
\arguments{
\item{object_list}{A list contains expression data with mapped gene symbol}

\item{gse_name}{A vector contains the name of the objects that you want to combine}
}
\value{
A SummarizedExperiment Object contains combined data from several objects
}
\description{
Combine samples with common genes from selected objects
}
