% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_curatedTBData.R
\name{CombineObjects}
\alias{CombineObjects}
\title{Combine samples with common genes from selected studies,
usually run after \code{\link{MatchProbe}}}
\usage{
CombineObjects(
  object_list,
  experiment_name = NULL,
  UseAssay = NULL,
  annotationColName = NULL,
  list_name = NULL,
  batch.adjust = TRUE,
  par.prior = TRUE,
  prior.plots = FALSE,
  mean.only = FALSE,
  ref.batch = NULL,
  BPPARAM = bpparam("SerialParam")
)
}
\arguments{
\item{object_list}{A list contains expression data with probes mapped to gene symbol.}

\item{experiment_name}{A character/vector to choose the name of the experiment from MultiAssayExperiment Object.}

\item{annotationColName}{A character/ indicates feature of interest in the object's column data.
This argument passes to `mod` parameter in \code{\link[sva]{ComBat}}. Default is NULL.}

\item{list_name}{A character/vector contains object name to be selected to merge.}

\item{batch.adjust}{A logical value indicating whether adjust for the batch effect.
Default is TRUE.}

\item{prior.plots}{Additional argument passed to \code{\link[sva]{ComBat}}.}

\item{mean.only}{Additional argument passed to \code{\link[sva]{ComBat}}.}

\item{ref.batch}{Additional argument passed to \code{\link[sva]{ComBat}}.}

\item{BPPARAM}{Additional argument passed to \code{\link[sva]{ComBat}}.}
}
\value{
A SummarizedExperiment Object contains combined data from several studies.
}
\description{
Combine samples with common genes from selected studies,
usually run after \code{\link{MatchProbe}}
}
\examples{
list_name <-  c("GSE101705","GSE54992","GSE19444")
data_list <-  get_curatedTBData(list_name)
object_norm <- lapply(data_list, function(x)
                               Normalization(x, microarray_method = "quantile",
                               RNAseq_method = "TMM", experiment_name = "assay_raw"))
object_match <- lapply(object_norm, function(x)
                               MatchProbe(x, UseAssay = c("TMM","quantile","RMA"),
                               createExperimentName = "assay_MatchProbe"))
sobject <- CombineObjects(object_match, list_name,
                          experiment_name = "assay_MatchProbe",
                          annotationColName = "TBStatus",
                         batch.adjust = TRUE)
}
