% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_curatedTBData.R
\name{create_ColData}
\alias{create_ColData}
\title{Create Column data information from Illumina HumanHT-12 V3.0/4.0 expression beadchip and Affymetrix Human Genome U133 Plus 2.0 Array}
\usage{
create_ColData(geo_access)
}
\arguments{
\item{geo_access}{A character that contains GEO accession number.}
}
\value{
A DataFrame object contains the patient information for each sample
}
\description{
Create Column data information from Illumina HumanHT-12 V3.0/4.0 expression beadchip and Affymetrix Human Genome U133 Plus 2.0 Array
}
\examples{
create_ColData("GSE39939")
}
