% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateMultiAssayObject.R
\name{Create_MultiAssay_object}
\alias{Create_MultiAssay_object}
\title{Create MultiAssay Experiment Object from Summarized Experiment Object.}
\usage{
Create_MultiAssay_object(sobject)
}
\arguments{
\item{sobject}{A summarized experiment object}
}
\value{
A MultiAssay Experiment Object contains 1. Input Summarized Experiment Object. 2. read count matrix collapse by gene symbol
}
\description{
Create MultiAssay Experiment Object from Summarized Experiment Object.
}
\examples{
dat("GSE39939_sobject")
Create_MultiAssay_object(GSE39939_sobject)
}
