% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTBData.R
\name{curatedTBData}
\alias{curatedTBData}
\title{Create curated Tuberculosis transcriptome data from GPL6947/GPL10558/GPL570}
\usage{
curatedTBData(geo_access, ...)
}
\description{
Create curated Tuberculosis transcriptome data from GPL6947/GPL10558/GPL570
}
