% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataSummary}
\alias{DataSummary}
\title{Summarized tuberculosis studies included in the curatedTBData}
\format{
A data frame with 49 rows and 14 variables
\describe{
  \item{GEOAccession}
  \item{Platform}
  \item{GeographicalRegion}
  \item{Tissue}
  \item{Age}
  \item{HIVStatus}
  \item{DiagnosisMethod}
  \item{Control}
  \item{LTBI}
  \item{PTB}
  \item{OD}
  \item{Total}
  \item{Notes}
  \item{GeneralType}
}
}
\source{

}
\usage{
DataSummary
}
\description{
A dataset containing metadata information of 49 tuberculosis studies.
The variables are as follows:
}
\keyword{datasets}
