% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_auc.R
\name{.check_input}
\alias{.check_input}
\title{Function to checking the input list format}
\usage{
.check_input(object_list)
}
\arguments{
\item{object_list}{A \code{list} of
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
}
\description{
Function to checking the input list format
}
