% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_auc.R
\name{.expand_study}
\alias{.expand_study}
\title{Expand study section for SignatureInfoTraining}
\usage{
.expand_study(GSE_sig)
}
\arguments{
\item{GSE_sig}{A `data.frame` contains information about each signature and
its training/discovery dataset(s) name. Default is `NULL`.}
}
\value{
A expand `data.frame` for gene signatures and dataset.
}
\description{
Expand study section for SignatureInfoTraining
}
