% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_AUC_CI.R
\name{.get_auc_stats}
\alias{.get_auc_stats}
\title{Obtain pvalue, emprirical AUC, and Bootstrap Confidence Interval for each signature.}
\usage{
.get_auc_stats(
  SE_scored,
  annotationColName = "TBStatus",
  signatureColNames,
  num.boot = NULL,
  percent = 0.95,
  AUC.abs = FALSE
)
}
\arguments{
\item{SE_scored}{A \code{SummarizedExperiment} object from TB signature profiling.}

\item{annotationColName}{A string specifying the feature of interest in the object's column data}

\item{signatureColNames}{A character/vector string contains name of gene signature.}

\item{num.boot}{Integer. Number of bootstrapping. If `NULL` (default), no confidence interval is computed.}

\item{percent}{Numeric. A number between 0 and 1, indicating the percentage of
confidence interval. Default is `0.95`.}

\item{AUC.abs}{Boolean. If AUC.abs = `TRUE`, return the AUC values from function \code{\link[ROCit]{rocit}}.
If AUC.abs = `FALSE`, return the AUC values for `max(AUC, 1-AUC)`.}
}
\value{
A data frame/datatable contains p-value from two-sample t-test and AUC value for each signature.
}
\description{
Obtain pvalue, emprirical AUC, and Bootstrap Confidence Interval for each signature.
}
