% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{get_auc_distribution}
\alias{get_auc_distribution}
\title{Obtain ridge plots for emprirical AUC distribution for signature scores.}
\usage{
get_auc_distribution(aucs_result)
}
\arguments{
\item{aucs_result_dat}{A dataframe contains AUC and p-value for certain TB signatures across different datasets}
}
\value{
Ridge plot with median line
}
\description{
Obtain ridge plots for emprirical AUC distribution for signature scores.
}
\examples{
aucs_result <- data.frame(Signature=c("Anderson_42", "Anderson_OD_51", "Berry_393"), AUC=rnorm(3))
p_ridge <- get_auc_distribution(aucs_result)
}
