% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{get_auc_distribution}
\alias{get_auc_distribution}
\title{Obtain ridge plots for emprirical AUC distribution for signature scores.}
\usage{
get_auc_distribution(aucs_result)
}
\arguments{
\item{aucs_result}{A dataframe contains signatures, p-value, and AUC, can be obtained from `combine_auc`.}
}
\value{
Ridge plot with median line
}
\description{
Obtain ridge plots for emprirical AUC distribution for signature scores.
}
\examples{
aucs_result <- data.frame(Signature=c("Anderson_42", "Anderson_OD_51", "Berry_393"), AUC=ruinf(3,0,0.5))
p_ridge <- get_auc_distribution(aucs_result)
}
