% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{get_auc_stats}
\alias{get_auc_stats}
\title{Obtain pvalue, emprirical AUC, and 95% CI for each signature using two-sample t-test, ROCit::rocit, and bootstraping}
\usage{
get_auc_stats(
  SE_scored,
  annotationColName = "TBStatus",
  signatureColNames,
  num.boot = NULL,
  percent = 0.95
)
}
\arguments{
\item{SE_scored}{A SummarizedExperiment Object from TB signature profiling.}

\item{annotationColName}{A character indicates feature of interest in the object's column data}

\item{signatureColNames}{A character/vector contains name of gene signature.}

\item{num.boot}{Number of bootstrapping.}
}
\value{
A data frame/datatable contains p-value from two-sample t-test and AUC value for each signature.
}
\description{
Obtain pvalue, emprirical AUC, and 95% CI for each signature using two-sample t-test, ROCit::rocit, and bootstraping
}
