% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_curatedTBData.R
\name{get_curatedTBData}
\alias{get_curatedTBData}
\title{Load curatedTBData into SummarizedExperiment/MultiAssayExperiment object.}
\usage{
get_curatedTBData(geo_access, include.SCAN = FALSE, include.reprocess = TRUE)
}
\arguments{
\item{geo_access}{A character/vector that contains geo accession number. If All, get all avaible studies.}

\item{include.SCAN}{A logical value indicates whether include normalized data processed by SCAN into the final output.
The default in FALSE. SCAN-normalized results are available for
Affymetrix Microarray: GSE31348, GSE36238, GSE41055, GSE54992, GSE73408
Agilent two-color Microarray: GSE25534}

\item{include.reprocess}{A logical value indicated whether include reprocessed RNA-seq data when available.
The default is TRUE. Reprocess results are available for
Illumina RNA-seq: GSE94438, GSE79362, GSE89403, GSE107991, GSE107992, GSE107993,
GSE107994, GSE101705, GSE107104, GSE112104}
}
\value{
A list of SummarizedExperiment/MultiAssayExperiment objects
}
\description{
\code{get_curatedTBData} loads available curatedTBData.
}
\examples{
get_curatedTBData(geo_access = "GSE39939")
get_curatedTBData(geo_access = c("GSE39939","GSE107993"))
get_curatedTBData(geo_access = "All")
}
