% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{get_pvalue_auc}
\alias{get_pvalue_auc}
\title{Obtain two-sample t-test pvalues and emprirical AUC for signature scores.}
\usage{
get_pvalue_auc(SE_scored, annotationColName = "Disease", signatureColNames)
}
\arguments{
\item{SE_scored}{A summarized experiment from TB signature profiling}

\item{annotationColName}{A character usually the column name of for TB status (default is "Disease")}

\item{signatureColNames}{A character vector that contains gene signature name}
}
\value{
A data contains p-value from two-sample t-test and AUC value for each signature
}
\description{
Obtain two-sample t-test pvalues and emprirical AUC for signature scores.
}
\examples{
sobject_TBSig <- get_sobject_TBSig(mobject,"PTB","Latent")
SE_scored <- runTBsigProfiler(input = sobject_TBSig, useAssay = "counts", signatures = TBsignatures, algorithm = "ssGSEA", combineSigAndAlgorithm = TRUE, parallel.sz = 1)
annotationColName <- "Disease"
signatureColNames <- names(colData(SE_scored))[-c(1:2)]
results <- get_pvalue_auc(SE_scored,"Disease",signatureColNames)
}
