% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchProbe.R
\name{MatchProbe}
\alias{MatchProbe}
\alias{MatchProbe,SummarizedExperiment-method}
\alias{MatchProbe,MultiAssayExperiment-method}
\title{S4 method matches probeID to gene symbol by creating MultiAssayExperiement object from either SummarizedExperiment or MultiAssayExperiment Object.}
\usage{
MatchProbe(theObject, ...)

\S4method{MatchProbe}{SummarizedExperiment}(
  theObject,
  UseAssay,
  experiment_type = "assay_MatchProbe",
  only.matrix = FALSE
)

\S4method{MatchProbe}{MultiAssayExperiment}(
  theObject,
  UseAssay,
  experiment_type = "assay_MatchProbe",
  only.matrix = FALSE
)
}
\arguments{
\item{theObject}{A SummarizedExperiment/MultiAssayExperiment Object.}

\item{...}{Extra named arguments passed to function.}

\item{UseAssay}{A character indicats the assay names (partial name) of the SummarizedExperiment Object.}

\item{experiment_type}{A character for the names of the matched output.}
}
\description{
S4 method matches probeID to gene symbol by creating MultiAssayExperiement object from either SummarizedExperiment or MultiAssayExperiment Object.
}
