% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchProbe.R
\name{MatchProbe}
\alias{MatchProbe}
\title{Match ProbeID to gene symbol by creating MultiAssayExperiement object from SummarizedExperiment/MultiAssayExperiment Object.}
\usage{
MatchProbe(theObject, ...)
}
\description{
Match ProbeID to gene symbol by creating MultiAssayExperiement object from SummarizedExperiment/MultiAssayExperiment Object.
}
