% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_curatedTBData.R
\docType{class}
\name{Mobject-class}
\alias{Mobject-class}
\alias{Mobject}
\title{Mobject Class
Create MultiAssayExperiment object for curatedTBData}
\description{
Mobject Class
Create MultiAssayExperiment object for curatedTBData
}
\section{Slots}{

\describe{
\item{\code{assay_reprocess}}{A matrix contatins gene expression data}

\item{\code{assay_raw}}{A matrix contatins gene expression data}

\item{\code{row_data}}{A DataFrame class contains gene expression data with different dimensions}

\item{\code{primary}}{A DataFrame class contains sample information}

\item{\code{meta_data}}{A MIAME class contains experiment information}
}}

