% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTBData.R
\docType{class}
\name{Mobject-class}
\alias{Mobject-class}
\alias{Mobject}
\title{Create MultiAssay Experiment object}
\description{
Create MultiAssay Experiment object
}
\section{Slots}{

\describe{
\item{\code{assay_reprocess}}{A matrix contatins gene expression data}

\item{\code{assay_raw}}{Another matrix contatins gene expression data}

\item{\code{row_data}}{A DataFrame contains gene expression data with different dimensions}

\item{\code{column_data}}{A DataFrame contains sample information}

\item{\code{meta_data}}{A MIAME class contains experiment information}
}}

