% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsetting.R, R/TBSignatureProfiling.R
\name{remove_empty_object}
\alias{remove_empty_object}
\title{Remove empty objects after single subsetting}
\usage{
remove_empty_object(k)

remove_empty_object(k)
}
\arguments{
\item{k}{A list of objects after single subsetting}

\item{multi_object}{A list contains both SummariexExperiment/MultiAssayExpriment objects}
}
\value{
A list contains non-empty SummariexExperiment/MultiAssayExpriment object
}
\description{
Remove empty objects after single subsetting

Remove empty objects from list contains both SummariexExperiment and MultiAssayExpriment objects
}
