% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{SignatureFilter}
\alias{SignatureFilter}
\alias{SignatureFilter,list,character-method}
\alias{SignatureFilter,SummarizedExperiment,character-method}
\title{Subset signatures scores and disease status from Coldata of SummarizedExperiment Objects.}
\usage{
SignatureFilter(sig_list, gset, ...)

\S4method{SignatureFilter}{list,character}(sig_list, gset, annotationColName = "TBStatus")

\S4method{SignatureFilter}{SummarizedExperiment,character}(sig_list, gset, GSE, annotationColName = "TBStatus")
}
\arguments{
\item{sig_list}{SummarizedExperiment object(s) produced from `TBSignatureProfiler::runTBsigProfiler`.
Names of each list should be the GEO accession of the study.}

\item{gset}{A character/vector contians name(s) of the signatures.}

\item{...}{Extra named arguments passed to function.}

\item{annotationColName}{A character indicates feature of interest in the object's column data.}

\item{GSE}{A character/vector for the GEO accession of the SummarizedExperiment Object.
Need GSE value when the input is a single object instead of the list.}
}
\description{
Subset signatures scores and disease status from Coldata of SummarizedExperiment Objects.
}
\seealso{
TBSignatureProfiler::TBSignatureProfiler.
}
