% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSignatureProfiling.R
\name{SignatureFilter}
\alias{SignatureFilter}
\title{Get signatures with only siganture scores and disease status}
\usage{
SignatureFilter(result_list, gset)
}
\arguments{
\item{result_list}{A list of SummarizedExperiment Object from `runTBSignatureProfiler`}

\item{gset}{A list of signatures want to select}
}
\value{
A list of DataFrame
}
\description{
Get signatures with only siganture scores and disease status
}
