% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTBData.R
\docType{class}
\name{Sobject}
\alias{Sobject}
\title{Create SummarizedExperiment object}
\description{
Create SummarizedExperiment object
}
\section{Slots}{

\describe{
\item{\code{assay}}{A matrix contatins gene expression data}

\item{\code{row_data}}{A DataFrame class contains gene information}

\item{\code{column_data}}{A DataFrame class contains sample information}

\item{\code{meta_data}}{A MIAME class contains experiment information}
}}

