% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_curatedTBData.R
\name{subset_curatedTBData}
\alias{subset_curatedTBData}
\alias{subset_curatedTBData,SummarizedExperiment-method}
\alias{subset_curatedTBData,MultiAssayExperiment-method}
\title{Subsetting curatedTBData based on single/multiple conditions}
\usage{
subset_curatedTBData(theObject, ...)

\S4method{subset_curatedTBData}{SummarizedExperiment}(
  theObject,
  annotationColName,
  annotationCondition,
  assayName = NULL,
  ...
)

\S4method{subset_curatedTBData}{MultiAssayExperiment}(
  theObject,
  annotationColName,
  annotationCondition,
  assayName = NULL,
  useAssay = NULL,
  ...
)
}
\arguments{
\item{theObject}{A [SummarizedExperiment][SummarizedExperiment::SummarizedExperiment-class]
/[MultiAssayExperiment][MultiAssayExperiment::MultiAssayExperiment-class] object.}

\item{...}{Extra named arguments passed to function.}

\item{annotationColName}{A character indicates feature of interest in the object's annotation data.}

\item{annotationCondition}{A vector of character indicates conditions want to be selected.}

\item{assayName}{A character indicates the name of the assay from the input object.
The default is `NULL`. When assayName is `NULL`, the function selects the first
assay along the assay list.}

\item{useAssay}{A character indicates the name of the assay when the
[SummarizedExperiment][SummarizedExperiment::SummarizedExperiment-class] is
selected from the [MultiAssayExperiment][MultiAssayExperiment::MultiAssayExperiment-class] object.
The default is `NULL`. When assayName is `NULL`, the function selects the first
assay along the assay list.}
}
\description{
\code{subset_curatedTBData} selects desired samples from curatedTBData
database based pre-specified conditions.
}
\examples{
obj <-  curatedTBData("GSE74092", dryrun = FALSE, curated.only = TRUE)
subset_curatedTBData(obj[[1]], annotationColName = "TBStatus",
                     annotationCondition = c("Control","PTB"))
}
