% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_curatedTBData.R
\name{subset_curatedTBData}
\alias{subset_curatedTBData}
\alias{subset_curatedTBData,SummarizedExperiment-method}
\alias{subset_curatedTBData,MultiAssayExperiment-method}
\title{Subsetting curatedTBData based on single/multiple conditions}
\usage{
subset_curatedTBData(theObject, ...)

\S4method{subset_curatedTBData}{SummarizedExperiment}(
  theObject,
  annotationColName,
  annotationCondition,
  assayName,
  ...
)

\S4method{subset_curatedTBData}{MultiAssayExperiment}(
  theObject,
  annotationColName,
  annotationCondition,
  assayName,
  ...
)
}
\arguments{
\item{theObject}{A [SummarizedExperiment][SummarizedExperiment::SummarizedExperiment-class]
/[MultiAssayExperiment][MultiAssayExperiment::MultiAssayExperiment-class] object.}

\item{...}{Extra named arguments passed to function.}

\item{annotationColName}{A character indicates feature of interest in the object's column data.}

\item{annotationCondition}{A vector indicates conditions want to be selected.}

\item{assayName}{A character indicates the name of the assay from the input object.}
}
\description{
\code{subset_curatedTBData} selects desired samples from curatedTBData
database according to pre-specified conditions.
}
