% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_curatedTBData.R
\name{subset_curatedTBData}
\alias{subset_curatedTBData}
\alias{subset_curatedTBData,SummarizedExperiment-method}
\alias{subset_curatedTBData,MultiAssayExperiment-method}
\title{Subsetting objects based on single/multiple conditions}
\usage{
subset_curatedTBData(theObject, ...)

\S4method{subset_curatedTBData}{SummarizedExperiment}(
  theObject,
  annotationColName,
  annotationCondition,
  UseAssay,
  ...
)

\S4method{subset_curatedTBData}{MultiAssayExperiment}(
  theObject,
  annotationColName,
  annotationCondition,
  UseAssay = NULL,
  experiment_name,
  ...
)
}
\arguments{
\item{theObject}{A SummarizedExperiment/MultiAssayExperiment object.}

\item{...}{Extra named arguments passed to function.}

\item{annotationColName}{A character indicates feature of interest in the object's column data.}

\item{annotationCondition}{A vector indicates conditions want to be subsetted.}

\item{UseAssay}{A character indicates the name of the assay (expression matrix) within the object.
Need this argument when the input is a SummarizedExperiment object.}

\item{experiment_name}{A character indicates the name of the experiment within MultiAssayExperiment object.
Expect theObject[[experiment_name]] to be a matrix. Two special cases are:
When experiment_name is "all". Perform whole MultiAssayExperiment subsetting, output is in the form of MultiAsaayExperiment object.
When experiment_name is "assay_raw". Perform subsetting on the SummarizedExperiment Object.}
}
\description{
Subsetting objects based on single/multiple conditions
}
